LayoutTests/inspector/unit-tests/object-utilities-expected.txt


== Running test suite: ObjectUtilities
-- Running test case: Object.shallowEqual
PASS: shallowEqual of empty object literals should be true.
PASS: shallowEqual of empty array literals should be true.
PASS: shallowEqual of empty array and object literals should be false.
PASS: shallowEqual of an object with itself should be true.
PASS: shallowEqual of equal objects should be true.
PASS: shallowEqual of equal objects should be true.
PASS: shallowEqual of unequal objects should be false.
PASS: shallowEqual of unequal objects should be false.
PASS: shallowEqual of an object and null should be false.
PASS: shallowEqual of an object and non-object should be false.
PASS: shallowEqual of a non-object with itself should be false.
PASS: shallowEqual of non-objects should be false.
PASS: shallowEqual of objects with equal constructors should be true.
PASS: shallowEqual of objects with different constructors should be false.
PASS: shallowEqual of objects with different constructors should be false.
